\name{mlegp-internal}
\alias{mlegp-internal}
\alias{calcVarMatrix}
\alias{calcCorOneObs}
\alias{calcLogLike}
\alias{calcLogLikeManual}
\alias{gpObs}
\alias{gpDim}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Internal Functions for Gaussian Processes }
\description{
  Internal Functions for Gaussian Processes
}
\details{
  These functions should not be called by the user
}
\references{ \url{http://www1.easternct.edu/dancikg/} }
\author{ Garrett M. Dancik \email{ dancikg@easternct.edu } }
\keyword{ internal }
